取りあえず動くInventory
