+++ Assets/Scripts/Inventory.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 3220bac874b3e4f2e8a7a51f646098f3
-timeCreated: 1442408525
+guid: 2847b617bcac04b8b914078047d9ae48
+timeCreated: 1442539170
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
