+++ Assets/Scripts/InventoryManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3220bac874b3e4f2e8a7a51f646098f3
+timeCreated: 1442408525
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
